translation_system/mcp_servers/excel_mcp/requirements.txt
# Excel MCP Server Dependencies

# MCP Framework
mcp>=0.9.0

# Excel Processing
openpyxl>=3.1.0
xlrd>=2.0.1
pandas>=2.0.0

# HTTP Client
requests>=2.31.0
aiohttp>=3.9.0
aiohttp-cors>=0.7.0

# JWT Token Validation
PyJWT>=2.8.0
cryptography>=41.0.0

# Language Detection
langdetect>=1.0.9

# Utilities
python-dateutil>=2.8.2
